Log the incoming URL
This demonstrates the problem:

The browser's development console shows that some (all?) of the static
things are 404s:

http://localhost:8080/static/sticky-footer.css e.g.

And examining the wsgi environment, you can see that the string "snurk"
doesn't appear anywhere, so it's clear that the python code couldn't
possibly generate the correct URLs.  So either nginx needs to give
pyramid more information, or nginx needs to somehow modify the responses
en passant (which sound like a terrible idea).